+++ src/modules/gacha/index.js
@@ -1,2 +1,3 @@
 export * from '../../asset-gacha.js';
 export * from './admin-validation.js';
+export * from './simulation.js';